+++ python_modules/dagster-graphql/dagster_graphql/implementation/fetch_partition_subsets.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 from dagster import _check as check
 from dagster._core.definitions.partitions.subset import PartitionsSubset
-from dagster._core.definitions.partitions.utils import CachingDynamicPartitionsLoader
+from dagster._core.instance.types import CachingDynamicPartitionsLoader
 from dagster._core.remote_representation.external_data import AssetNodeSnap
 from dagster._core.storage.partition_status_cache import get_partition_subsets
 from dagster._core.workspace.context import WorkspaceRequestContext
