+++ python_modules/dagster/dagster/_core/definitions/partitions/definition/dynamic.py
@@ -1,6 +1,6 @@
 from collections.abc import Sequence
 from datetime import datetime
-from typing import Callable, NamedTuple, Optional, Union
+from typing import TYPE_CHECKING, Callable, NamedTuple, Optional, Union
 
 import dagster._check as check
 from dagster._annotations import PublicAttr, deprecated_param, public
@@ -17,8 +17,10 @@ from dagster._core.definitions.partitions.definition.partitions_definition impor
 )
 from dagster._core.definitions.partitions.partition import Partition
 from dagster._core.errors import DagsterInvalidDefinitionError
-from dagster._core.instance import DynamicPartitionsStore
 from dagster._core.types.pagination import PaginatedResults
+
+if TYPE_CHECKING:
+    from dagster._core.instance import DynamicPartitionsStore
 
 
 @deprecated_param(
@@ -123,8 +125,8 @@ class DynamicPartitionsDefinition(
             return super().__str__()
 
     def _ensure_dynamic_partitions_store(
-        self, dynamic_partitions_store: Optional[DynamicPartitionsStore]
-    ) -> DynamicPartitionsStore:
+        self, dynamic_partitions_store: Optional["DynamicPartitionsStore"]
+    ) -> "DynamicPartitionsStore":
         if dynamic_partitions_store is None:
             check.failed(
                 "The instance is not available to load partitions. You may be seeing this error"
@@ -139,7 +141,7 @@ class DynamicPartitionsDefinition(
     def get_partition_keys(
         self,
         current_time: Optional[datetime] = None,
-        dynamic_partitions_store: Optional[DynamicPartitionsStore] = None,
+        dynamic_partitions_store: Optional["DynamicPartitionsStore"] = None,
     ) -> Sequence[str]:
         """Returns a list of strings representing the partition keys of the
         PartitionsDefinition.
@@ -168,7 +170,7 @@ class DynamicPartitionsDefinition(
                 ).get_dynamic_partitions(partitions_def_name=self._validated_name())
 
     def get_serializable_unique_identifier(
-        self, dynamic_partitions_store: Optional[DynamicPartitionsStore] = None
+        self, dynamic_partitions_store: Optional["DynamicPartitionsStore"] = None
     ) -> str:
         with partition_loading_context(dynamic_partitions_store=dynamic_partitions_store) as ctx:
             return self._ensure_dynamic_partitions_store(
@@ -192,7 +194,7 @@ class DynamicPartitionsDefinition(
         self,
         partition_key: str,
         current_time: Optional[datetime] = None,
-        dynamic_partitions_store: Optional[DynamicPartitionsStore] = None,
+        dynamic_partitions_store: Optional["DynamicPartitionsStore"] = None,
     ) -> bool:
         with partition_loading_context(current_time, dynamic_partitions_store) as ctx:
             if self.partition_fn:
