+++ python_modules/dagster/dagster/_core/definitions/partitions/utils/__init__.py
@@ -3,9 +3,6 @@ from dagster._core.definitions.partitions.utils.base import (
     raise_error_on_duplicate_partition_keys as raise_error_on_duplicate_partition_keys,
     raise_error_on_invalid_partition_key_substring as raise_error_on_invalid_partition_key_substring,
 )
-from dagster._core.definitions.partitions.utils.dynamic import (
-    CachingDynamicPartitionsLoader as CachingDynamicPartitionsLoader,
-)
 from dagster._core.definitions.partitions.utils.mapping import (
     get_builtin_partition_mapping_types as get_builtin_partition_mapping_types,
     infer_partition_mapping as infer_partition_mapping,
