+++ python_modules/dagster/dagster/_core/definitions/schedule_definition.py
@@ -957,7 +957,7 @@ class ScheduleDefinition(IHasInternalInit):
             ScheduleExecutionData: Contains list of run requests, or skip message if present.
 
         """
-        from dagster._core.definitions.partitions.utils import CachingDynamicPartitionsLoader
+        from dagster._core.instance.types import CachingDynamicPartitionsLoader
 
         check.inst_param(context, "context", ScheduleEvaluationContext)
         execution_fn: Callable[..., ScheduleEvaluationFunctionReturn]
