+++ python_modules/dagster/dagster/_core/workspace/context.py
@@ -17,7 +17,6 @@ from dagster._core.definitions.asset_key import AssetKey
 from dagster._core.definitions.assets.graph.remote_asset_graph import RemoteRepositoryAssetNode
 from dagster._core.definitions.data_time import CachingDataTimeResolver
 from dagster._core.definitions.data_version import CachingStaleStatusResolver
-from dagster._core.definitions.partitions.utils import CachingDynamicPartitionsLoader
 from dagster._core.definitions.selector import (
     JobSelector,
     JobSubsetSelector,
@@ -28,6 +27,7 @@ from dagster._core.definitions.selector import (
 from dagster._core.errors import DagsterCodeLocationLoadError, DagsterCodeLocationNotFoundError
 from dagster._core.execution.plan.state import KnownExecutionState
 from dagster._core.instance import DagsterInstance
+from dagster._core.instance.types import CachingDynamicPartitionsLoader
 from dagster._core.loader import LoadingContext
 from dagster._core.remote_representation import (
     CodeLocation,
